Improved typing, refactor
