+++ protostar/commands/test/cheatcodes/reflect/misc.py
@@ -2,7 +2,7 @@ from typing import Union, Type
 from starkware.cairo.lang.vm.relocatable import RelocatableValue
 from starkware.cairo.lang.vm.vm_consts import VmConstsReference
 
-from protostar.commands.test.cheatcodes.reflect.reflect_cairo_struct import CairoStruct
+from protostar.commands.test.cheatcodes.reflect.cairo_struct import CairoStruct
 
 ReflectInputType = Union[VmConstsReference, RelocatableValue, None, int]
 ReflectValueType = Union[CairoStruct, RelocatableValue, int]
