+++ protostar/commands/test/environments/test_execution_environment.py
@@ -21,7 +21,7 @@ from protostar.starknet.cheatcode import Cheatcode
 from protostar.starknet.execution_environment import ExecutionEnvironment
 from protostar.utils.abi import has_function_parameters
 from protostar.utils.hook import Hook
-from protostar.commands.test.cheatcodes.reflect.reflect_cairo_struct import (
+from protostar.commands.test.cheatcodes.reflect.cairo_struct import (
     CairoStructHintLocal,
 )
